requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
attrdict==2.0.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
blinker==1.4
braintree==3.29.0
celery==3.1.23
click==6.6
configparser==3.5.0
decorator==4.0.10
Flask==0.11.1
Flask-Mail==0.9.1
gunicorn==19.7.1
ipdb==0.8
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
kombu==3.0.35
MarkupSafe==0.23
passlib==1.6.5
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
ptyprocess==0.5.1
py2neo==2.0.8
Pygments==2.1.3
pymongo==3.3.0
python-premailer==0.5.7
pytz==2016.6.1
PyYAML==3.11
redis==2.10.5
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
Werkzeug==0.11.10
GitPython==2.1.3
markdown2==2.3.4
